update - modularize the network api service layer - App layer doesn't have Retrofit dependency anymore. All handled in the Network module
