+++ app/src/main/java/com/neudesic/myapplication/data/repository/DadJokeRepository.kt
@@ -1,11 +1,11 @@
 package com.neudesic.myapplication.data.repository
 
+import com.neudesic.module.core.network.DadJokeAPIService
 import com.neudesic.module.core.network.NetworkError
 import com.neudesic.module.core.network.NetworkException
 import com.neudesic.module.core.network.NetworkSuccess
 import com.neudesic.myapplication.domain.model.DadJoke
 import com.neudesic.myapplication.domain.model.DataResult
-import com.neudesic.myapplication.domain.network.DadJokeAPIService
 import com.neudesic.myapplication.domain.repository.DadJokeRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
@@ -26,7 +26,7 @@ class DadJokeRepositoryImpl(private val dadJokeAPIService: DadJokeAPIService) : 
                     result.message = response.e.message ?: "Network Error"
                 }
                 // TODO: look into mapper from DTO to Domain
-                is NetworkSuccess -> result.data = response.data.toDomain()
+                is NetworkSuccess -> result.data = DadJoke.fromDTO(response.data)
             }
             return@withContext result
         }
