+++ app/src/main/java/com/neudesic/myapplication/di/ApplicationModule.kt
@@ -1,7 +1,7 @@
 package com.neudesic.myapplication.di
 
+import com.neudesic.module.core.network.DadJokeAPIService
 import com.neudesic.myapplication.data.repository.DadJokeRepositoryImpl
-import com.neudesic.myapplication.domain.network.DadJokeAPIService
 import com.neudesic.myapplication.domain.repository.DadJokeRepository
 import dagger.Module
 import dagger.Provides
@@ -12,7 +12,6 @@ import javax.inject.Singleton
 @InstallIn(SingletonComponent::class)
 @Module
 object ApplicationModule {
-    // TODO: try @binds?
     @Provides
     @Singleton
     fun providesDadJokeRepo(dadJokeAPIService: DadJokeAPIService): DadJokeRepository {
