+++ module/core/network/README.md
@@ -1,0 +1,11 @@
+# Internal Module - Network
+
+Handles all things related to the App's http calls
+This module integrates with [Retrofit](https://square.github.io/retrofit/) and exposes the following API
+
+- retrofitClient - retrofitClient with `GSON`converter and custom callAdapter
+- NetworkResult - Convenient interface for network call result
+
+## Usage
+
+TODO
