+++ module/core/network/src/main/java/com/neudesic/module/core/network/CoreNetworkModule.kt
@@ -1,7 +1,9 @@
 package com.neudesic.module.core.network
 
+import okhttp3.OkHttpClient
 import retrofit2.CallAdapter
+import retrofit2.Retrofit
 
 interface CoreNetworkModule {
-    fun networkResultCallAdapterFactory(): CallAdapter.Factory
+    fun getDadJokeAPIService(baseUrl: String): DadJokeAPIService
 }
